+++ prefab_cloud_python/client.py
@@ -5,8 +5,10 @@ from .feature_flag_client import FeatureFlagClient
 from .context_shape_aggregator import ContextShapeAggregator
 from .log_path_aggregator import LogPathAggregator
 from .logger_client import LoggerClient
+from .logger_filter import LoggerFilter
 from .options import Options
 from typing import Optional
+import base64
 import prefab_pb2 as Prefab
 import uuid
 import requests
@@ -126,4 +128,10 @@ class Client:
             headers=headers,
             data=body.SerializeToString(),
             auth=("authuser", self.options.api_key or ""),
+
+    def logging_filter(self):
+        return LoggerFilter(
+            self.config_client(),
+            prefix=self.options.log_prefix,
+            log_boundary=self.options.log_boundary
         )