+++ prefab_cloud_python/config_client.py
@@ -3,6 +3,7 @@ from .config_resolver import ConfigResolver
 from .read_write_lock import ReadWriteLock
 from .config_value_unwrapper import ConfigValueUnwrapper
 from .context import Context
+from google.protobuf.json_format import MessageToJson, Parse
 
 import grpc
 import threading
@@ -11,6 +12,10 @@ import sseclient
 import base64
 import prefab_pb2 as Prefab
 import functools
+import os
+
+
+STALE_CACHE_WARN_HOURS = 5
 
 
 class InitializationTimeoutException(Exception):
@@ -52,8 +57,12 @@ class ConfigClient:
             "debug", "Initialize ConfigClient: acquired write lock"
         )
 
+        self.set_cache_path()
+
         if self.options.is_local_only():
             self.finish_init("local only")
+        elif self.options.has_datafile():
+            self.load_json_file(self.options.datafile)
         else:
             self.load_checkpoint()
             self.start_checkpointing_thread()
@@ -68,7 +77,10 @@ class ConfigClient:
             self.base_client.context_shape_aggregator.push(Context(context))
 
         if value is not None:
-            return ConfigValueUnwrapper.unwrap(value, key, context)
+            raw_config = self.config_resolver.raw(key)
+            return ConfigValueUnwrapper.deepest_value(
+                value, raw_config, self.config_resolver, context
+            ).unwrap()
         else:
             return self.handle_default(key, default)
 
@@ -97,6 +109,8 @@ class ConfigClient:
 
     def load_checkpoint(self):
         if self.load_checkpoint_from_api_cdn():
+            return
+        if self.load_cache():
             return
         self.base_client.logger.log_internal("warn", "No success loading checkpoints")
 
@@ -149,7 +163,7 @@ class ConfigClient:
         else:
             self.base_client.logger.log_internal(
                 "info",
-                f"Checkpoint remote_cdn_api failed to load. Response {response.status_code}",
+                f"Checkpoint remote_cdn_api failed to load: {e.__class__.__name__}",
             )
             return False
 
@@ -158,6 +172,17 @@ class ConfigClient:
         project_env_id = configs.config_service_pointer.project_env_id
         self.config_resolver.project_env_id = project_env_id
         starting_highwater_mark = self.config_loader.highwater_mark
+
+        default_contexts = {}
+        if configs.default_context and configs.default_context.contexts is not None:
+            for context in configs.default_context.contexts:
+                values = {}
+                for k, v in context.values.items():
+                    values[k] = ConfigValueUnwrapper(v, self.config_resolver).unwrap()
+                default_contexts[context.type] = values
+
+        self.config_resolver.default_context = default_contexts
+
         for config in configs.configs:
             self.config_loader.set(config, source)
         if self.config_loader.highwater_mark > starting_highwater_mark:
@@ -173,12 +198,64 @@ class ConfigClient:
         self.config_resolver.update()
         self.finish_init(source)
 
+    def cache_configs(self, configs):
+        if not self.options.use_local_cache:
+            return
+        with open(self.cache_path, "w") as f:
+            f.write(MessageToJson(configs))
+            self.base_client.logger.log_internal(
+                "debug", f"Cached configs to {self.cache_path}"
+            )
+
+    def load_cache(self):
+        if not self.options.use_local_cache:
+            return False
+        try:
+            with open(self.cache_path, "r") as f:
+                configs = Parse(f.read(), Prefab.Configs())
+                self.load_configs(configs, "cache")
+
+                hours_old = round(
+                    (time.mktime(time.localtime()) - os.path.getmtime(self.cache_path))
+                    / 3600,
+                    2,
+                )
+                if hours_old > STALE_CACHE_WARN_HOURS:
+                    self.base_client.logger.log_internal(
+                        "info", f"Stale Cache Load: {hours_old} hours old"
+                    )
+                return True
+        except OSError as e:
+            self.base_client.logger.log_internal("info", e)
+            return False
+
+    def load_json_file(self, datafile):
+        with open(datafile) as f:
+            configs = Parse(f.read(), Prefab.Configs())
+            self.load_configs(configs, "datafile")
+
     def finish_init(self, source):
         if not self.init_lock._write_locked:
             return
         self.base_client.logger.log_internal("info", f"Unlocked config via {source}")
         self.init_lock.release_write()
         self.base_client.logger.set_config_client(self)
+
+    def set_cache_path(self):
+        dir = os.environ.get(
+            "XDG_CACHE_HOME", os.path.join(os.environ["HOME"], ".cache")
+        )
+        file_name = f"prefab.cache.{self.base_client.options.api_key_id}.json"
+        self.cache_path = os.path.join(dir, file_name)
+
+    @property
+    def cache_path(self):
+        os.makedirs(os.path.dirname(self._cache_path), exist_ok=True)
+        return self._cache_path
+
+    @cache_path.setter
+    def cache_path(self, path):
+        self._cache_path = path
 
     @functools.cache
     def grpc_channel(self):